tests: split mocked tests into unit (client) and server_mocked (MCP tools); add markers to pyproject
