+++ tests/test_flink_sql_gateway_client.py
@@ -134,71 +134,4 @@ async def test_close_operation_mocked() -> None:
 
 
 # MCP Server tests using in-memory testing approach
-@pytest.mark.asyncio
-async def test_mcp_server_open_session(client) -> None:
-    """Test opening a session through MCP server"""
-    result = await client.call_tool("open_new_session", {})
-
-    data = result.data
-    assert "sessionHandle" in data
-    assert isinstance(data["sessionHandle"], str)
-
-
-@pytest.mark.asyncio
-async def test_mcp_server_get_config(client, session_handle) -> None:
-    """Test getting session config through MCP server"""
-    # Get config
-    config_result = await client.call_tool(
-        "get_config", {"session_handle": session_handle}
-    )
-
-    config_data = config_result.data
-    assert isinstance(config_data, dict)
-
-
-@pytest.mark.asyncio
-async def test_mcp_server_configure_session(client, session_handle) -> None:
-    """Test configuring a session through MCP server"""
-    # The configure_session call may succeed with an empty response
-    configure_result = await client.call_tool(
-        "configure_session",
-        {
-            "session_handle": session_handle,
-            "statement": "SET execution.runtime-mode = 'batch'",
-        },
-    )
-    # configure_session returns an empty dict {} which becomes None in fastmcp
-    # The important thing is that it doesn't error
-    assert not configure_result.is_error
-
-
-@pytest.mark.asyncio
-async def test_mcp_server_fetch_result_page(client, session_handle) -> None:
-    """Test pagination via the fetch_result_page tool."""
-    # Start a streaming query to obtain an operation handle
-    start = await client.call_tool(
-        "run_query_stream_start",
-        {"session_handle": session_handle, "query": "SELECT 1"},
-    )
-    start_data = start.data
-    op = start_data.get("operationHandle")
-    assert isinstance(op, str) and op
-
-    # Fetch page 0 and assert it is not EOS
-    page0 = await client.call_tool(
-        "fetch_result_page",
-        {"session_handle": session_handle, "operation_handle": op, "token": 0},
-    )
-    p0 = page0.data
-    assert p0.get("isEnd") is False
-    assert p0.get("nextToken") == 1
-    assert "page" in p0 and isinstance(p0["page"], dict)
-
-    # Fetch page 1 and assert EOS
-    page1 = await client.call_tool(
-        "fetch_result_page",
-        {"session_handle": session_handle, "operation_handle": op, "token": 1},
-    )
-    p1 = page1.data
-    assert p1.get("isEnd") is True
-    assert p1.get("nextToken") == 2
+pytestmark = [pytest.mark.asyncio, pytest.mark.unit]
